+++ fiddles/c/fiddle-0021-Functions/functions/src/main.c
@@ -1,0 +1,14 @@
+#include <stdio.h>
+
+void printMessage (void) {
+    printf ( "\n" );
+    printf ( "{{ ʕ・ɭ・ʔ }}" );
+    printf ( "\n" );
+}
+
+int main(void) {
+
+    printMessage ();    
+
+    return 0;
+}